# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SmartLecture AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  avatar: string | null;
  /** @nullable */
  bio?: string | null;
  level: number;
  xp: number;
  streak: number;
  joinedAt?: string;
}

export interface AuthResult {
  token: string;
  user: UserProfile;
}

export interface UserProfileUpdate {
  name?: string;
  bio?: string;
  avatar?: string;
}

export interface DayActivity {
  day: string;
  minutes: number;
  xp: number;
}

export interface UserStats {
  xp: number;
  level: number;
  xpToNextLevel: number;
  streak: number;
  totalStudyMinutes: number;
  lecturesCount: number;
  flashcardsReviewed: number;
  quizzesCompleted: number;
  weeklyActivity: DayActivity[];
}

export interface SubjectProgress {
  subject: string;
  color: string;
  lecturesTotal: number;
  lecturesCompleted: number;
  avgQuizScore: number;
}

export interface MonthlyStudy {
  month: string;
  hours: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  title: string;
  description: string;
  xpEarned: number;
  createdAt: string;
}

export interface UserProgress {
  subjectProgress: SubjectProgress[];
  monthlyStudy: MonthlyStudy[];
  recentActivity: ActivityItem[];
}

export interface Achievement {
  id: number;
  title: string;
  description: string;
  icon: string;
  category: string;
  xpReward: number;
  unlocked: boolean;
  /** @nullable */
  unlockedAt?: string | null;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  progressMax?: number | null;
}

export type LectureStatus = typeof LectureStatus[keyof typeof LectureStatus];


export const LectureStatus = {
  processing: 'processing',
  ready: 'ready',
  error: 'error',
} as const;

export interface Lecture {
  id: number;
  title: string;
  subject: string;
  subjectColor: string;
  duration: number;
  status: LectureStatus;
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  tags: string[];
  viewCount?: number;
  createdAt: string;
}

export type LectureDetail = Lecture & ({
  /** @nullable */
  summary?: string | null;
  keyPoints?: string[];
  /** @nullable */
  flashcardDeckId?: number | null;
  /** @nullable */
  quizId?: number | null;
});

export interface LectureInput {
  title: string;
  subject: string;
  tags?: string[];
  transcript?: string;
}

export interface LectureUpdate {
  title?: string;
  subject?: string;
  tags?: string[];
}

export interface Concept {
  term: string;
  definition: string;
}

export interface LectureSummary {
  lectureId: number;
  summary: string;
  keyPoints: string[];
  concepts: Concept[];
}

export interface MindMapNode {
  id: string;
  label: string;
  level: number;
  color: string;
  /** @nullable */
  description?: string | null;
}

export interface MindMapEdge {
  from: string;
  to: string;
}

export interface MindMap {
  lectureId: number;
  nodes: MindMapNode[];
  edges: MindMapEdge[];
}

export interface Subject {
  name: string;
  color: string;
  icon: string;
  lectureCount: number;
}

export interface FlashcardDeck {
  id: number;
  title: string;
  subject: string;
  subjectColor: string;
  cardCount: number;
  masteredCount: number;
  dueCount: number;
  /** @nullable */
  lastStudied: string | null;
  createdAt: string;
}

export interface Flashcard {
  id: number;
  deckId: number;
  front: string;
  back: string;
  mastered: boolean;
  /** @nullable */
  difficulty?: string | null;
}

export type FlashcardDeckDetail = FlashcardDeck & {
  cards: Flashcard[];
};

export interface FlashcardDeckInput {
  title: string;
  subject: string;
  lectureId?: number;
}

export interface FlashcardInput {
  front: string;
  back: string;
}

export interface StudySessionInput {
  cardsReviewed: number;
  correctCount: number;
  durationSeconds: number;
}

export interface XpUpdate {
  xpEarned: number;
  newTotalXp: number;
  newLevel: number;
  streakUpdated: boolean;
  newStreak?: number;
}

export interface Quiz {
  id: number;
  title: string;
  subject: string;
  subjectColor: string;
  questionCount: number;
  /** @nullable */
  bestScore: number | null;
  attempts: number;
  createdAt: string;
}

export interface Question {
  id: number;
  quizId: number;
  text: string;
  options: string[];
  correctIndex: number;
  /** @nullable */
  explanation?: string | null;
}

export type QuizDetail = Quiz & {
  questions: Question[];
};

export interface QuizInput {
  title: string;
  subject: string;
  lectureId?: number;
}

export interface QuizAttemptInput {
  answers: number[];
}

export interface QuestionResult {
  questionId: number;
  correct: boolean;
  selectedIndex: number;
  correctIndex: number;
  /** @nullable */
  explanation?: string | null;
}

export interface QuizResult {
  score: number;
  correctCount: number;
  totalCount: number;
  xpEarned: number;
  passed: boolean;
  questionResults: QuestionResult[];
}

export interface ChatMessageInput {
  message: string;
  /** @nullable */
  lectureId?: number | null;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface SearchResults {
  lectures: Lecture[];
  flashcardDecks: FlashcardDeck[];
  quizzes: Quiz[];
}

export type GetLecturesParams = {
subject?: string;
search?: string;
};

export type GlobalSearchParams = {
q: string;
};