# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartLecture AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /lectures:
    get:
      operationId: getLectures
      tags: [lectures]
      summary: List all lectures
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lecture list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lecture"
    post:
      operationId: createLecture
      tags: [lectures]
      summary: Create a lecture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LectureInput"
      responses:
        "201":
          description: Created lecture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
...
  /lectures/{id}:
    get:
      operationId: getLecture
      tags: [lectures]
      summary: Get a single lecture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lecture detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LectureDetail"
    patch:
      operationId: updateLecture
      tags: [lectures]
      summary: Update a lecture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LectureUpdate"
      responses:
        "200":
          description: Updated lecture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
    delete:
      operationId: deleteLecture
      tags: [lectures]
      summary: Delete a lecture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /lectures/{id}/summary:
    get:
      operationId: getLectureSummary
      tags: [lectures]
      summary: Get AI-generated summary for a lecture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LectureSummary"
  /lectures/{id}/mindmap:
    get:
      operationId: getLectureMindmap
      tags: [lectures]
      summary: Get mind map data for a lecture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mind map nodes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MindMap"
  /flashcard-decks:
    get:
      operationId: getFlashcardDecks
      tags: [flashcards]
      summary: List all flashcard decks
      responses:
        "200":
          description: Deck list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlashcardDeck"
    post:
      operationId: createFlashcardDeck
      tags: [flashcards]
      summary: Create a flashcard deck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardDeckInput"
      responses:
        "201":
          description: Created deck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardDeck"
  /flashcard-decks/{id}:
    get:
      operationId: getFlashcardDeck
      tags: [flashcards]
      summary: Get a flashcard deck with its cards
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deck with cards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardDeckDetail"
  /flashcard-decks/{id}/cards:
    post:
      operationId: addFlashcard
      tags: [flashcards]
      summary: Add a card to a deck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardInput"
      responses:
        "201":
          description: Created card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flashcard"
  /flashcard-decks/{id}/study:
    post:
      operationId: recordFlashcardStudy
      tags: [flashcards]
      summary: Record a study session result for a deck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudySessionInput"
      responses:
        "200":
          description: XP and streak update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/XpUpdate"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      required: [id, name, email, avatar, level, xp, streak]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        level:
          type: integer
        xp:
          type: integer
        streak:
          type: integer
        joinedAt:
          type: string
...
    Lecture:
      type: object
      required: [id, title, subject, subjectColor, duration, status, createdAt, tags]
      properties:
        id:
          type: integer
        title:
          type: string
        subject:
          type: string
        subjectColor:
          type: string
        duration:
          type: integer
        status:
          type: string
          enum: [processing, ready, error]
        transcript:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        viewCount:
          type: integer
        createdAt:
          type: string
    LectureDetail:
      allOf:
        - $ref: "#/components/schemas/Lecture"
        - type: object
          properties:
            summary:
              type: ["string", "null"]
            keyPoints:
              type: array
              items:
                type: string
            flashcardDeckId:
              type: ["integer", "null"]
            quizId:
              type: ["integer", "null"]
    LectureInput:
      type: object
      required: [title, subject]
      properties:
        title:
          type: string
        subject:
          type: string
        tags:
          type: array
          items:
            type: string
        transcript:
          type: string
    LectureUpdate:
      type: object
      properties:
        title:
          type: string
        subject:
          type: string
        tags:
          type: array
          items:
            type: string
    LectureSummary:
      type: object
      required: [lectureId, summary, keyPoints, concepts]
      properties:
        lectureId:
          type: integer
        summary:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        concepts:
          type: array
          items:
            $ref: "#/components/schemas/Concept"
    Concept:
      type: object
      required: [term, definition]
      properties:
        term:
          type: string
        definition:
          type: string
    MindMap:
      type: object
      required: [lectureId, nodes, edges]
      properties:
        lectureId:
          type: integer
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/MindMapNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/MindMapEdge"
    MindMapNode:
      type: object
      required: [id, label, level, color]
      properties:
        id:
          type: string
        label:
          type: string
        level:
          type: integer
        color:
          type: string
        description:
          type: ["string", "null"]
    MindMapEdge:
      type: object
      required: [from, to]
      properties:
        from:
          type: string
        to:
          type: string
...
    FlashcardDeck:
      type: object
      required: [id, title, subject, subjectColor, cardCount, masteredCount, dueCount, lastStudied, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        subject:
          type: string
        subjectColor:
          type: string
        cardCount:
          type: integer
        masteredCount:
          type: integer
        dueCount:
          type: integer
        lastStudied:
          type: ["string", "null"]
        createdAt:
          type: string
    FlashcardDeckDetail:
      allOf:
        - $ref: "#/components/schemas/FlashcardDeck"
        - type: object
          required: [cards]
          properties:
            cards:
              type: array
              items:
                $ref: "#/components/schemas/Flashcard"
    FlashcardDeckInput:
      type: object
      required: [title, subject]
      properties:
        title:
          type: string
        subject:
          type: string
        lectureId:
          type: integer
    Flashcard:
      type: object
      required: [id, deckId, front, back, mastered]
      properties:
        id:
          type: integer
        deckId:
          type: integer
        front:
          type: string
        back:
          type: string
        mastered:
          type: boolean
        difficulty:
          type: ["string", "null"]
    FlashcardInput:
      type: object
      required: [front, back]
      properties:
        front:
          type: string
        back:
          type: string
    StudySessionInput:
      type: object
      required: [cardsReviewed, correctCount, durationSeconds]
      properties:
        cardsReviewed:
          type: integer
        correctCount:
          type: integer
        durationSeconds:
          type: integer
    XpUpdate:
      type: object
      required: [xpEarned, newTotalXp, newLevel, streakUpdated]
      properties:
        xpEarned:
          type: integer
        newTotalXp:
          type: integer
        newLevel:
          type: integer
        streakUpdated:
          type: boolean
        newStreak:
          type: integer
...