# lib/api-zod/src/generated/types/flashcard.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SmartLecture AI API
 * OpenAPI spec version: 0.1.0
 */

export interface Flashcard {
  id: number;
  deckId: number;
  front: string;
  back: string;
  mastered: boolean;
  /** @nullable */
  difficulty?: string | null;
}